# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentOS API",
    "description": "AgentOS gateway API contract.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "PlatformTenant": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenant_id",
          "display_name",
          "status",
          "created_at",
          "enabled_modules"
        ],
        "properties": {
          "tenant_id": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "enabled_modules": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PlatformReadiness": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenant_id",
          "capability_count",
          "capability_statuses",
          "connector_count",
          "connector_statuses",
          "owner_input_count",
          "owner_input_statuses",
          "workspace_status",
          "residency_status"
        ],
        "properties": {
          "tenant_id": {
            "type": "string"
          },
          "capability_count": {
            "type": [
              "null",
              "integer"
            ]
          },
          "capability_statuses": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "connector_count": {
            "type": [
              "null",
              "integer"
            ]
          },
          "connector_statuses": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "owner_input_count": {
            "type": [
              "null",
              "integer"
            ]
          },
          "owner_input_statuses": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "workspace_status": {
            "type": [
              "null",
              "string"
            ]
          },
          "residency_status": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PlatformUsage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenant_id",
          "display_name",
          "usage_day",
          "domain",
          "model",
          "currency",
          "cost_recorded",
          "record_count",
          "input_tokens_total",
          "output_tokens_total",
          "cached_tokens_total",
          "tokens_total",
          "cost_total",
          "monthly_token_budget"
        ],
        "properties": {
          "tenant_id": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "usage_day": {
            "type": "string"
          },
          "domain": {
            "type": [
              "null",
              "string"
            ]
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "cost_recorded": {
            "type": "boolean"
          },
          "record_count": {
            "type": "integer"
          },
          "input_tokens_total": {
            "type": "integer"
          },
          "output_tokens_total": {
            "type": "integer"
          },
          "cached_tokens_total": {
            "type": "integer"
          },
          "tokens_total": {
            "type": "integer"
          },
          "cost_total": {
            "type": [
              "null",
              "string"
            ]
          },
          "monthly_token_budget": {
            "type": [
              "null",
              "integer"
            ]
          }
        }
      },
      "PlatformProvider": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "provider",
          "configured",
          "mode"
        ],
        "properties": {
          "provider": {
            "type": "string"
          },
          "configured": {
            "type": "boolean"
          },
          "mode": {
            "type": "string"
          }
        }
      },
      "PlatformLlmProvider": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "provider_id",
          "display_name",
          "base_url",
          "reasoning_model",
          "fast_model",
          "timeout_ms",
          "structured_mode",
          "status",
          "is_default",
          "secret_configured",
          "config_version",
          "updated_at",
          "last_probe",
          "secret_fingerprint",
          "secret_last4"
        ],
        "properties": {
          "provider_id": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "base_url": {
            "type": "string"
          },
          "reasoning_model": {
            "type": "string"
          },
          "fast_model": {
            "type": "string"
          },
          "timeout_ms": {
            "type": "integer",
            "minimum": 1000,
            "maximum": 120000
          },
          "structured_mode": {
            "type": "string",
            "enum": [
              "json_object",
              "json_schema"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "CONFIGURED",
              "VERIFIED",
              "FAILED"
            ]
          },
          "is_default": {
            "type": "boolean"
          },
          "secret_configured": {
            "type": "boolean"
          },
          "config_version": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_probe": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": false,
            "required": [
              "outcome",
              "latency_ms",
              "http_status",
              "error_class",
              "probed_at"
            ],
            "properties": {
              "outcome": {
                "type": "string",
                "enum": [
                  "PASS",
                  "FAIL"
                ]
              },
              "latency_ms": {
                "type": [
                  "null",
                  "integer"
                ]
              },
              "http_status": {
                "type": [
                  "null",
                  "integer"
                ]
              },
              "error_class": {
                "type": [
                  "null",
                  "string"
                ]
              },
              "probed_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "secret_fingerprint": {
            "type": [
              "null",
              "string"
            ]
          },
          "secret_last4": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PlatformTenantsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformTenant"
            }
          }
        }
      },
      "PlatformUsageResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformUsage"
            }
          }
        }
      },
      "PlatformProvidersResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "providers"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformProvider"
            }
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformLlmProvider"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error_code",
          "message",
          "retryable",
          "correlation_id"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "correlation_id": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/platform/tenants": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenantsResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/tenants/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTenant"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/tenants/{id}/readiness": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformReadiness"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/usage": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "from",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "to",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformUsageResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/providers": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformProvidersResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}